Renamespace Task/Type/Type (#667)
