+++ src/Services/TaskOutputJoiner/Factory.php
@@ -1,7 +1,7 @@
 <?php
 namespace App\Services\TaskOutputJoiner;
 
-use App\Entity\Task\Type\Type as TaskType;
+use App\Entity\Task\Type as TaskType;
 
 class Factory
 {
