+++ src/Services/TaskOutputJoiner/LinkIntegrityTaskOutputJoiner.php
@@ -2,7 +2,7 @@
 namespace App\Services\TaskOutputJoiner;
 
 use App\Entity\Task\Output as TaskOutput;
-use App\Entity\Task\Type\Type as TaskType;
+use App\Entity\Task\Type as TaskType;
 use App\Services\TaskTypeService;
 use webignition\InternetMediaType\InternetMediaType;
 use webignition\InternetMediaType\Parser\Parser as InternetMediaTypeParser;
