+++ src/Services/TaskOutputJoiner/TaskOutputJoinerInterface.php
@@ -2,7 +2,7 @@
 namespace App\Services\TaskOutputJoiner;
 
 use App\Entity\Task\Output as TaskOutput;
-use App\Entity\Task\Type\Type as TaskType;
+use App\Entity\Task\Type as TaskType;
 
 interface TaskOutputJoinerInterface
 {
