+++ src/Services/TaskPreProcessor/LinkIntegrityTaskPreProcessor.php
@@ -7,7 +7,7 @@ use GuzzleHttp\Psr7\Request;
 use Psr\Http\Message\UriInterface;
 use Psr\Log\LoggerInterface;
 use App\Entity\Task\Task;
-use App\Entity\Task\Type\Type;
+use App\Entity\Task\Type;
 use App\Repository\TaskRepository;
 use App\Services\HttpClientService;
 use App\Services\StateService;
