+++ src/Services/TaskTypeDomainsToIgnoreService.php
@@ -1,7 +1,7 @@
 <?php
 namespace App\Services;
 
-use App\Entity\Task\Type\Type as TaskType;
+use App\Entity\Task\Type as TaskType;
 
 class TaskTypeDomainsToIgnoreService
 {
