+++ tests/Functional/Services/CrawlJobUrlCollectorTest.php
@@ -6,7 +6,7 @@ use App\Entity\CrawlJobContainer;
 use App\Entity\Job\Job;
 use App\Entity\State;
 use App\Entity\Task\Task;
-use App\Entity\Task\Type\Type;
+use App\Entity\Task\Type;
 use App\Services\CrawlJobContainerService;
 use App\Services\CrawlJobUrlCollector;
 use App\Services\StateService;
