+++ tests/Functional/Services/TaskPostProcessor/UrlDiscoveryTaskPostProcessorTest.php
@@ -4,7 +4,7 @@ namespace App\Tests\Functional\Services\TaskPostProcessor;
 
 use App\Entity\Job\Ammendment;
 use App\Entity\Task\Task;
-use App\Entity\Task\Type\Type;
+use App\Entity\Task\Type;
 use App\Services\CrawlJobContainerService;
 use App\Services\StateService;
 use App\Services\TaskPostProcessor\UrlDiscoveryTaskPostProcessor;
