+++ tests/Unit/Entity/Task/TaskTest.php
@@ -6,7 +6,7 @@ use ReflectionClass;
 use App\Entity\State;
 use App\Entity\Task\Output;
 use App\Entity\Task\Task;
-use App\Entity\Task\Type\Type;
+use App\Entity\Task\Type;
 use App\Entity\TimePeriod;
 use App\Tests\Factory\ModelFactory;
